Label positioning fix
Fixed the focused/filled label positioning by basing the size of the text off a font size animation instead of a scale.  The scale was causing us to have to calculate exaxtly how far each label based on width needed to be translated, but this seems to just work.

Also changed some of the positioning to better fit material, please check me there.

Finally, the label was wrapped in a View such that pointer events can be ignored.  Previously, this was causing taps to be ignored on the label text and focus would not occur.