248 SheetBottom Flashing on first animation
Removed the idential .ios and .android files and just used a .native file (Sorry it's hard to see the diff).  Adjusted everything to be based on full height so that the initial value isn't 0 for a tenth of a second before the animation starts.  This seems to fix both native and web.