Added prop to use input casing for buttons
By default in Material, buttons should be capitalized, but this gives the user the option to use input casing in case their designs have text in a button that is not capitalized.  Children could also be used here, but this solution uses MB's own rendering to accomplish it which seems easiser to using children.